requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
requests>=2.28.0
